fix(lwql): address CodeRabbit review findings on the LWQL skill/CLI/dashboard surface
Nine substantive findings from PR review, verified against source before
fixing: a query-cache key missing projectId, positional args replaced with a
named-parameter object, a cross-field grid-placement bound now enforced by a
Zod refine, three call sites of a hardcoded granularity tuple index unified
behind a shared lwqlGranularityStepSchema, a wrong parameter-missing
remediation message, an openapi integer/number type mismatch traced back to
its Zod source, a missing status-200 assertion, a granularity picker that
showed no checkmark when running at its implicit default, and a CLI
verb-refusal test loop missing its list/get attempts.

Also fixes two bugs surfaced by validation, not part of the original
findings: an invalid grid-placement test fixture (gridColumn 1 + colSpan 2
exceeds the 2-column bound the refine above added) and a mock-call-count
assertion that didn't account for runChartCommand reading through the same
service.get the get verb itself calls.

+++ platform/app/src/app/api/analytics-sql/[[...route]]/app.charts.v1.ts
@@ -156,10 +156,10 @@ const chartSchema = z.object({
   platformUrl: z.string(),
   /** `null` when the chart has never been placed, or has been unplaced. */
   dashboardId: z.string().nullable(),
-  gridColumn: z.number(),
-  gridRow: z.number(),
-  colSpan: z.number(),
-  rowSpan: z.number(),
+  gridColumn: z.number().int(),
+  gridRow: z.number().int(),
+  colSpan: z.number().int(),
+  rowSpan: z.number().int(),
 });
 
 const chartListSchema = z.object({ data: z.array(chartSchema) });
